move transformer fron nlp to common
