+++ nemo/collections/common/parts/transformer/transformer_modules.py
@@ -15,11 +15,9 @@
 # limitations under the License.
 
 import math
-from dataclasses import dataclass
 
 import numpy as np
 import torch
-from omegaconf.omegaconf import MISSING
 from torch import nn
 from torch.nn.functional import gelu
